fix: refine not found handling

+++ src/controllers/banner.controller.ts
@@ -2,7 +2,7 @@ import { Response } from 'express';
 import { IPaginationQuery, IReqUser } from '../utils/interfaces';
 import response from '../utils/response';
 import BannerModel, { bannerDAO, TypeBanner } from '../models/banner.model';
-import { FilterQuery } from 'mongoose';
+import { FilterQuery, isValidObjectId } from 'mongoose';
 
 export default {
 	async create(req: IReqUser, res: Response) {
@@ -57,6 +57,9 @@ export default {
 	async findOne(req: IReqUser, res: Response) {
 		try {
 			const { id } = req.params;
+			if (!isValidObjectId(id)) {
+				return response.notfound(res, 'banner not found');
+			}
 
 			const result = await BannerModel.findById(id);
 			if (!result) {
@@ -71,6 +74,9 @@ export default {
 	async update(req: IReqUser, res: Response) {
 		try {
 			const { id } = req.params;
+			if (!isValidObjectId(id)) {
+				return response.notfound(res, 'banner not found');
+			}
 
 			const result = await BannerModel.findByIdAndUpdate(id, req.body, {
 				new: true,
@@ -87,6 +93,9 @@ export default {
 	async remove(req: IReqUser, res: Response) {
 		try {
 			const { id } = req.params;
+			if (!isValidObjectId(id)) {
+				return response.notfound(res, 'banner not found');
+			}
 
 			const result = await BannerModel.findByIdAndDelete(id, {
 				new: true,
