+++ src/controllers/category.controller.ts
@@ -2,6 +2,7 @@ import { Response } from 'express';
 import { IPaginationQuery, IReqUser } from '../utils/interfaces';
 import CategoryModel, { categoryDAO } from '../models/category.model';
 import response from '../utils/response';
+import { isValidObjectId } from 'mongoose';
 
 export default {
 	async create(req: IReqUser, res: Response) {
@@ -60,6 +61,10 @@ export default {
 	async findOne(req: IReqUser, res: Response) {
 		try {
 			const { id } = req.params;
+			if (!isValidObjectId(id)) {
+				return response.notfound(res, 'Category not found');
+			}
+
 			const result = await CategoryModel.findById(id);
 
 			if (!result) {
@@ -74,6 +79,10 @@ export default {
 	async update(req: IReqUser, res: Response) {
 		try {
 			const { id } = req.params;
+			if (!isValidObjectId(id)) {
+				return response.notfound(res, 'Category not found');
+			}
+
 			const result = await CategoryModel.findByIdAndUpdate(id, req.body, {
 				new: true,
 			});
@@ -89,6 +98,10 @@ export default {
 	async remove(req: IReqUser, res: Response) {
 		try {
 			const { id } = req.params;
+			if (!isValidObjectId(id)) {
+				return response.notfound(res, 'Category not found');
+			}
+
 			const result = await CategoryModel.findByIdAndDelete(id, { new: true });
 			if (!result) {
 				return response.notfound(res, 'Category not found');
