+++ src/controllers/event.controller.ts
@@ -2,7 +2,7 @@ import { Response } from 'express';
 import { IPaginationQuery, IReqUser } from '../utils/interfaces';
 import response from '../utils/response';
 import EventModel, { eventDAO, TEvent } from '../models/event.model';
-import { FilterQuery } from 'mongoose';
+import { FilterQuery, isValidObjectId } from 'mongoose';
 
 export default {
 	async create(req: IReqUser, res: Response) {
@@ -62,10 +62,13 @@ export default {
 	async findOne(req: IReqUser, res: Response) {
 		try {
 			const { id } = req.params;
+			if (!isValidObjectId(id)) {
+				return response.notfound(res, 'event not found');
+			}
 
 			const result = await EventModel.findById(id);
 			if (!result) {
-				response.notfound(res, 'Event not found');
+				response.notfound(res, 'event not found');
 			}
 
 			response.success(res, result, 'success find one event');
@@ -76,6 +79,9 @@ export default {
 	async update(req: IReqUser, res: Response) {
 		try {
 			const { id } = req.params;
+			if (!isValidObjectId(id)) {
+				return response.notfound(res, 'event not found');
+			}
 
 			const result = await EventModel.findByIdAndUpdate(id, req.body, {
 				new: true,
@@ -92,6 +98,9 @@ export default {
 	async remove(req: IReqUser, res: Response) {
 		try {
 			const { id } = req.params;
+			if (!isValidObjectId(id)) {
+				return response.notfound(res, 'event not found');
+			}
 
 			const result = await EventModel.findByIdAndDelete(id, {
 				new: true,
