+++ src/controllers/ticket.controller.ts
@@ -59,7 +59,7 @@ export default {
 		try {
 			const { eventId } = req.params;
 			if (!isValidObjectId(eventId)) {
-				return response.error(res, null, 'tickets not found');
+				return response.notfound(res, 'ticket not found');
 			}
 
 			const result = await TicketModel.find({ events: eventId }).exec();
@@ -71,10 +71,12 @@ export default {
 	async findOne(req: IReqUser, res: Response) {
 		try {
 			const { id } = req.params;
-
+			if (!isValidObjectId(id)) {
+				return response.notfound(res, 'ticket not found');
+			}
 			const result = await TicketModel.findById(id);
 			if (!result) {
-				response.notfound(res, 'Ticket not found');
+				response.notfound(res, 'ticket not found');
 			}
 
 			response.success(res, result, 'success find one ticket');
@@ -85,6 +87,9 @@ export default {
 	async update(req: IReqUser, res: Response) {
 		try {
 			const { id } = req.params;
+			if (!isValidObjectId(id)) {
+				return response.notfound(res, 'ticket not found');
+			}
 
 			const result = await TicketModel.findByIdAndUpdate(id, req.body, {
 				new: true,
@@ -100,6 +105,9 @@ export default {
 	async remove(req: IReqUser, res: Response) {
 		try {
 			const { id } = req.params;
+			if (!isValidObjectId(id)) {
+				return response.notfound(res, 'ticket not found');
+			}
 
 			const result = await TicketModel.findByIdAndDelete(id, {
 				new: true,
